migrate to slnx
