--- ZLinq.sln
@@ -1,96 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.13.35806.99
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{DDB4EF12-CC4E-5E88-E6FC-71A802F4DE97}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZLinq.Tests", "tests\ZLinq.Tests\ZLinq.Tests.csproj", "{F389F5ED-3C02-41A6-857A-344E5BE5C064}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{F462254E-52F2-4CA2-8518-583E60F81452}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZLinq", "src\ZLinq\ZLinq.csproj", "{58657F19-047A-45C7-A9F9-776772A6F664}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZLinq.__Unity", "src\ZLinq.__Unity\ZLinq.__Unity.csproj", "{7CC3CED3-98E8-4A1A-98CE-F47AEB3B4548}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sandbox", "sandbox", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Benchmark", "sandbox\Benchmark\Benchmark.csproj", "{5EFC61C2-13D0-4BC8-B475-183DFA2A3352}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleApp", "sandbox\ConsoleApp\ConsoleApp.csproj", "{358BF506-1640-4ADD-91FF-0444011675FC}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZLinq.FileSystem", "src\ZLinq.FileSystem\ZLinq.FileSystem.csproj", "{31ACFF78-0320-4143-B222-C3906BF85293}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZLinq.Json", "src\ZLinq.Json\ZLinq.Json.csproj", "{1B36B6F5-A573-406D-981A-7CCBA652E617}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileGen", "sandbox\FileGen\FileGen.csproj", "{873D634E-DA03-411C-AFED-F2651E7B749B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZLinq.Godot", "src\ZLinq.Godot\ZLinq.Godot.csproj", "{672A654D-F698-489F-BC10-7FF0A831D65D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZLinq.DropInGenerator", "src\ZLinq.DropInGenerator\ZLinq.DropInGenerator.csproj", "{D7E3A89C-7FF5-475A-86E0-5C3C70CAFD3F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F389F5ED-3C02-41A6-857A-344E5BE5C064}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F389F5ED-3C02-41A6-857A-344E5BE5C064}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F389F5ED-3C02-41A6-857A-344E5BE5C064}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F389F5ED-3C02-41A6-857A-344E5BE5C064}.Release|Any CPU.Build.0 = Release|Any CPU
-		{58657F19-047A-45C7-A9F9-776772A6F664}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{58657F19-047A-45C7-A9F9-776772A6F664}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{58657F19-047A-45C7-A9F9-776772A6F664}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{58657F19-047A-45C7-A9F9-776772A6F664}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7CC3CED3-98E8-4A1A-98CE-F47AEB3B4548}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7CC3CED3-98E8-4A1A-98CE-F47AEB3B4548}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7CC3CED3-98E8-4A1A-98CE-F47AEB3B4548}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7CC3CED3-98E8-4A1A-98CE-F47AEB3B4548}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5EFC61C2-13D0-4BC8-B475-183DFA2A3352}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5EFC61C2-13D0-4BC8-B475-183DFA2A3352}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5EFC61C2-13D0-4BC8-B475-183DFA2A3352}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5EFC61C2-13D0-4BC8-B475-183DFA2A3352}.Release|Any CPU.Build.0 = Release|Any CPU
-		{358BF506-1640-4ADD-91FF-0444011675FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{358BF506-1640-4ADD-91FF-0444011675FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{358BF506-1640-4ADD-91FF-0444011675FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{358BF506-1640-4ADD-91FF-0444011675FC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{31ACFF78-0320-4143-B222-C3906BF85293}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{31ACFF78-0320-4143-B222-C3906BF85293}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{31ACFF78-0320-4143-B222-C3906BF85293}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{31ACFF78-0320-4143-B222-C3906BF85293}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1B36B6F5-A573-406D-981A-7CCBA652E617}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1B36B6F5-A573-406D-981A-7CCBA652E617}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1B36B6F5-A573-406D-981A-7CCBA652E617}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1B36B6F5-A573-406D-981A-7CCBA652E617}.Release|Any CPU.Build.0 = Release|Any CPU
-		{873D634E-DA03-411C-AFED-F2651E7B749B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{873D634E-DA03-411C-AFED-F2651E7B749B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{873D634E-DA03-411C-AFED-F2651E7B749B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{873D634E-DA03-411C-AFED-F2651E7B749B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{672A654D-F698-489F-BC10-7FF0A831D65D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{672A654D-F698-489F-BC10-7FF0A831D65D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{672A654D-F698-489F-BC10-7FF0A831D65D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{672A654D-F698-489F-BC10-7FF0A831D65D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D7E3A89C-7FF5-475A-86E0-5C3C70CAFD3F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D7E3A89C-7FF5-475A-86E0-5C3C70CAFD3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D7E3A89C-7FF5-475A-86E0-5C3C70CAFD3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D7E3A89C-7FF5-475A-86E0-5C3C70CAFD3F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{F389F5ED-3C02-41A6-857A-344E5BE5C064} = {DDB4EF12-CC4E-5E88-E6FC-71A802F4DE97}
-		{58657F19-047A-45C7-A9F9-776772A6F664} = {F462254E-52F2-4CA2-8518-583E60F81452}
-		{7CC3CED3-98E8-4A1A-98CE-F47AEB3B4548} = {F462254E-52F2-4CA2-8518-583E60F81452}
-		{5EFC61C2-13D0-4BC8-B475-183DFA2A3352} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
-		{358BF506-1640-4ADD-91FF-0444011675FC} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
-		{31ACFF78-0320-4143-B222-C3906BF85293} = {F462254E-52F2-4CA2-8518-583E60F81452}
-		{1B36B6F5-A573-406D-981A-7CCBA652E617} = {F462254E-52F2-4CA2-8518-583E60F81452}
-		{873D634E-DA03-411C-AFED-F2651E7B749B} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
-		{672A654D-F698-489F-BC10-7FF0A831D65D} = {F462254E-52F2-4CA2-8518-583E60F81452}
-		{D7E3A89C-7FF5-475A-86E0-5C3C70CAFD3F} = {F462254E-52F2-4CA2-8518-583E60F81452}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2FFC585D-0FDF-489B-87D0-AB1122FD8C51}
-	EndGlobalSection
-EndGlobal
