+++ ZLinq.slnx
@@ -1,0 +1,18 @@
+<Solution>
+  <Folder Name="/sandbox/">
+    <Project Path="sandbox/Benchmark/Benchmark.csproj" />
+    <Project Path="sandbox/ConsoleApp/ConsoleApp.csproj" />
+  </Folder>
+  <Folder Name="/src/">
+    <Project Path="src/ZLinq.DropInGenerator/ZLinq.DropInGenerator.csproj" />
+    <Project Path="src/ZLinq.FileSystem/ZLinq.FileSystem.csproj" />
+    <Project Path="src/ZLinq.Godot/ZLinq.Godot.csproj" />
+    <Project Path="src/ZLinq.Json/ZLinq.Json.csproj" />
+    <Project Path="src/ZLinq.__Unity/ZLinq.__Unity.csproj" />
+    <Project Path="src/ZLinq/ZLinq.csproj" />
+    <Project Path="src/FileGen/FileGen.csproj" />
+  </Folder>
+  <Folder Name="/tests/">
+    <Project Path="tests/ZLinq.Tests/ZLinq.Tests.csproj" />
+  </Folder>
+</Solution>
