+++ src/FileGen/DropIn/IEnumerable.cs
