+++ src/FileGen/DropIn/Memory.cs
