+++ src/FileGen/DropIn/ReadOnlySpan.cs
