+++ src/FileGen/DropIn/Span.cs
